+++ sketch/src/androidTest/java/com/github/panpf/sketch/test/stateimage/DefaultErrorRulesTest.kt
@@ -20,9 +20,8 @@ import android.graphics.drawable.ColorDrawable
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.panpf.sketch.request.DisplayRequest
 import com.github.panpf.sketch.stateimage.ColorStateImage
-import com.github.panpf.sketch.stateimage.ErrorStateImage.DefaultMatcher
+import com.github.panpf.sketch.stateimage.ErrorStateImage.DefaultErrorRules
 import com.github.panpf.sketch.test.utils.TestAssets
-import com.github.panpf.sketch.test.utils.getTestContext
 import com.github.panpf.sketch.test.utils.getTestContextAndNewSketch
 import com.github.panpf.sketch.util.asOrThrow
 import org.junit.Assert
@@ -30,54 +29,48 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class ErrorStateImageDefaultMatcherTest {
-
-    @Test
-    fun testMatch() {
-        val context = getTestContext()
-        val request = DisplayRequest(context, TestAssets.SAMPLE_JPEG_URI)
-
-        DefaultMatcher(ColorStateImage(Color.RED)).apply {
-            Assert.assertTrue(match(request, null))
-        }
-
-        DefaultMatcher(ColorStateImage(Color.GREEN)).apply {
-            Assert.assertTrue(match(request, null))
-        }
-    }
+class DefaultErrorRulesTest {
 
     @Test
     fun testGetDrawable() {
         val (context, sketch) = getTestContextAndNewSketch()
         val request = DisplayRequest(context, TestAssets.SAMPLE_JPEG_URI)
 
-        DefaultMatcher(ColorStateImage(Color.RED)).apply {
+        DefaultErrorRules(ColorStateImage(Color.RED)).apply {
+            Assert.assertNotNull(getDrawable(sketch, request, null).getOrNull())
+        }
+
+        DefaultErrorRules(ColorStateImage(Color.GREEN)).apply {
+            Assert.assertNotNull(getDrawable(sketch, request, null).getOrNull())
+        }
+
+        DefaultErrorRules(ColorStateImage(Color.RED)).apply {
             Assert.assertEquals(
                 Color.RED,
-                getDrawable(sketch, request, null)!!.asOrThrow<ColorDrawable>().color
+                getDrawable(sketch, request, null).getOrNull()!!.asOrThrow<ColorDrawable>().color
             )
         }
 
-        DefaultMatcher(ColorStateImage(Color.GREEN)).apply {
+        DefaultErrorRules(ColorStateImage(Color.GREEN)).apply {
             Assert.assertEquals(
                 Color.GREEN,
-                getDrawable(sketch, request, null)!!.asOrThrow<ColorDrawable>().color
+                getDrawable(sketch, request, null).getOrNull()!!.asOrThrow<ColorDrawable>().color
             )
         }
     }
 
     @Test
     fun testToString() {
-        DefaultMatcher(ColorStateImage(Color.RED)).apply {
+        DefaultErrorRules(ColorStateImage(Color.RED)).apply {
             Assert.assertEquals(
-                "DefaultMatcher(ColorStateImage(IntColor(${Color.RED})))",
+                "DefaultErrorRules(ColorStateImage(IntColor(${Color.RED})))",
                 toString()
             )
         }
 
-        DefaultMatcher(ColorStateImage(Color.GREEN)).apply {
+        DefaultErrorRules(ColorStateImage(Color.GREEN)).apply {
             Assert.assertEquals(
-                "DefaultMatcher(ColorStateImage(IntColor(${Color.GREEN})))",
+                "DefaultErrorRules(ColorStateImage(IntColor(${Color.GREEN})))",
                 toString()
             )
         }
@@ -85,10 +78,10 @@ class ErrorStateImageDefaultMatcherTest {
 
     @Test
     fun testEqualsAndHashCode() {
-        val element1 = DefaultMatcher(ColorStateImage(Color.RED))
-        val element11 = DefaultMatcher(ColorStateImage(Color.RED))
-        val element2 = DefaultMatcher(ColorStateImage(Color.GREEN))
-        val element3 = DefaultMatcher(ColorStateImage(Color.BLUE))
+        val element1 = DefaultErrorRules(ColorStateImage(Color.RED))
+        val element11 = DefaultErrorRules(ColorStateImage(Color.RED))
+        val element2 = DefaultErrorRules(ColorStateImage(Color.GREEN))
+        val element3 = DefaultErrorRules(ColorStateImage(Color.BLUE))
 
         Assert.assertNotSame(element1, element11)
         Assert.assertNotSame(element1, element2)